add cil to mips visitor
